module_4_negotiation_logic/requirements.txt
cvxpy>=1.3.0
numpy>=1.21.0
pandas>=1.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
python-dotenv>=0.19.0
pydantic>=2.0.0
